+++ js/lib/definitions/mistral.js
@@ -27,13 +27,13 @@ export default class MistralDefinition extends Definition {
 
   get spec() {
     return _.assign(super.spec, {
-      WORKFLOWS_BLOCK: /^\s*workflows:\s*$/,
+      WORKFLOWS_BLOCK: /^(\s*)workflows:\s*$/,
       WORKFLOW_NAME: /^(\s*)([\w.-]+):\s*$/,
-      TASK_BLOCK: /^\s*tasks:\s*$/,
-      TASK_ACTION: /(.*)(action:\s+['"]*)([\w\s.]+)/,
-      SUCCESS_BLOCK: /^\s*on-success:\s*$/,
-      ERROR_BLOCK: /^\s*on-error:\s*$/,
-      COMPLETE_BLOCK: /^\s*on-complete:\s*$/,
+      TASK_BLOCK: /^(\s*)tasks:\s*$/,
+      TASK_ACTION: /(.*)(action:\s+['"]*)([\w.]+)/,
+      SUCCESS_BLOCK: /^(\s*)on-success:\s*$/,
+      ERROR_BLOCK: /^(\s*)on-error:\s*$/,
+      COMPLETE_BLOCK: /^(\s*)on-complete:\s*$/,
       TRANSITION: /^(\s*-\s*)(\w+)/
     });
   }
@@ -43,9 +43,9 @@ export default class MistralDefinition extends Definition {
       task: (starter, indent) => _.template(starter + '${name}:\n' + indent + 'action: ${ref}\n'),
       block: {
         base: () => _.template(`---\nversion: '2.0'\n\nworkflows:\n`),
-        workflow: (external='', internal=external + '  ') =>
+        workflow: (external, internal) =>
           _.template(external + '${name}:\n' + internal + 'type: ${type}\n'),
-        tasks: (indent='') => _.template(indent + 'tasks:\n'),
+        tasks: (indent) => _.template(indent + 'tasks:\n'),
         success: (indent) => _.template(indent + 'on-success:\n'),
         error: (indent) => _.template(indent + 'on-error:\n'),
         complete: (indent) => _.template(indent + 'on-complete:\n')
@@ -56,6 +56,18 @@ export default class MistralDefinition extends Definition {
 
   parseLine(state, line, lineNum) {
     state.workflowBlock = state.workflowBlock || new Sector();
+
+    let match = line.match(this.spec.WS_INDENT)[0];
+    if (match) {
+      if (state.unit) {
+        // if it is not the same character or if the prefix has any other chracters
+        if (state.unit !== match[0] || match.split(match[0]).join('') !== '') {
+          console.warn('Mixing tabs and spaces as indentation. Parser state in undefined.');
+        }
+      } else {
+        state.unit = match[0];
+      }
+    }
 
     if (this.spec.EMPTY_LINE.test(line)) {
       return;
@@ -126,6 +138,7 @@ export default class MistralDefinition extends Definition {
       const name = state.potentialWorkflow.getProperty('name');
       state.currentWorkflow = this.model.workflow(name, state.potentialWorkflow);
       state.taskBlock = state.currentWorkflow.getSector('taskBlock');
+      state.indent = indent;
     }
 
     state.potentialWorkflow = null;
@@ -136,8 +149,10 @@ export default class MistralDefinition extends Definition {
       let block = this.block('isTaskBlock', this.spec.TASK_BLOCK);
 
       if (block.enter(line, lineNum, state)) {
-        state.currentWorkflow.startSector('taskBlock', lineNum, 0);
-        state.currentWorkflow.endSector('taskBlock', lineNum + 1, 0);
+        const sector = state.currentWorkflow.getSector('taskBlock');
+        sector.setStart(lineNum, 0);
+        sector.setEnd(lineNum + 1, 0);
+        sector.indent = state.unit.repeat(state.isTaskBlock - 1);
         return;
       }
 
@@ -149,55 +164,27 @@ export default class MistralDefinition extends Definition {
     if (state.currentWorkflow && state.isTaskBlock) {
       state.currentWorkflow.endSector('taskBlock', lineNum + 1, 0);
 
-      let match;
-
-      match = this.spec.WORKFLOW_NAME.exec(line);
-      if (match) {
-        const [,starter,name] = match
-            , coords = [lineNum, starter.length, lineNum, (starter+name).length]
-            ;
-
-        if (!state.currentTask || starter.length === state.currentTask.starter.length) {
-          if (state.currentTask) {
-            state.currentTask.endSector('task', lineNum, 0);
-          }
-
-          const taskSector = new Sector(lineNum, 0).setType('task')
-              , nameSector = new Sector(...coords).setType('name')
-              ;
-
-          state.currentTask = this.model.task(name, {})
-            .setSector('task', taskSector)
-            .setSector('name', nameSector)
-            ;
-
-          taskSector.task = state.currentTask;
-
-          state.currentTask.starter = starter;
-
-          const TYPES = ['success', 'error', 'complete'];
-
-          _.each(TYPES, (type) => {
-            const sector = new Sector().setType(type);
-            sector.indent = starter + '  ';
-            sector.childStarter = starter + '    - ';
-
-            state.currentTask
-              .setProperty(type, [])
-              .setSector(type, sector)
-              ;
-          });
-
-          _.remove(state.untouchedTasks, (e) => e === name);
-        }
-      }
-
       if (state.currentTask) {
 
-        let handler;
+        let handler
+          , match
+          ;
 
         handler = this.handler('ref', this.spec.TASK_ACTION);
-        if (handler(line, lineNum, state.currentTask)) {
+        match = handler(line, lineNum, state.currentTask);
+        if (match) {
+          let [,_prefix] = match;
+
+          if (state.currentTask.isEmpty()) {
+            if (state.currentTask.starter === _prefix) {
+              state.currentTask.indent = state.unit.repeat(_prefix.length);
+            } else {
+              state.currentTask.starter += _prefix;
+            }
+          } else {
+            state.currentTask.indent = _prefix;
+          }
+
           return;
         }
 
@@ -208,7 +195,7 @@ export default class MistralDefinition extends Definition {
         if (block.enter(line, lineNum, state)) {
           const sector = state.currentTask.getSector('success');
           sector.setStart(lineNum, 0);
-          sector.indent = ' '.repeat(state.isSuccessBlock - 1);
+          sector.indent = state.unit.repeat(state.isSuccessBlock - 1);
           return;
         }
 
@@ -248,7 +235,7 @@ export default class MistralDefinition extends Definition {
         if (block.enter(line, lineNum, state)) {
           const sector = state.currentTask.getSector('error');
           sector.setStart(lineNum, 0);
-          sector.indent = ' '.repeat(state.isErrorBlock - 1);
+          sector.indent = state.unit.repeat(state.isErrorBlock - 1);
           return;
         }
 
@@ -288,7 +275,7 @@ export default class MistralDefinition extends Definition {
         if (block.enter(line, lineNum, state)) {
           const sector = state.currentTask.getSector('complete');
           sector.setStart(lineNum, 0);
-          sector.indent = ' '.repeat(state.isCompleteBlock - 1);
+          sector.indent = state.unit.repeat(state.isCompleteBlock - 1);
           return;
         }
 
@@ -323,6 +310,53 @@ export default class MistralDefinition extends Definition {
           }
         }
 
+      }
+
+      let match;
+
+      match = this.spec.WORKFLOW_NAME.exec(line);
+      if (match) {
+        const [,starter,name] = match
+            , coords = [lineNum, starter.length, lineNum, (starter+name).length]
+            ;
+
+        if (!state.currentTask || starter.length === state.currentTask.starter.length) {
+          if (state.currentTask) {
+            state.currentTask.endSector('task', lineNum, 0);
+          }
+
+          const taskSector = new Sector(lineNum, 0).setType('task')
+              , nameSector = new Sector(...coords).setType('name')
+              ;
+
+          state.currentTask = this.model.task(name, {})
+            .setSector('task', taskSector)
+            .setSector('name', nameSector)
+            ;
+
+          taskSector.task = state.currentTask;
+
+          state.currentTask.starter = starter;
+
+          const TYPES = ['success', 'error', 'complete'];
+
+          _.each(TYPES, (type) => {
+            const sector = new Sector().setType(type)
+                , outdent = state.taskBlock.indent
+                , unit = state.unit.repeat(starter.length - outdent.length)
+                ;
+
+            sector.indent = starter + unit;
+            sector.childStarter = sector.indent + unit + '- ';
+
+            state.currentTask
+              .setProperty(type, [])
+              .setSector(type, sector)
+              ;
+          });
+
+          _.remove(state.untouchedTasks, (e) => e === name);
+        }
       }
 
     }