+++ CHANGELOG.md
@@ -4,6 +4,9 @@ Changelog
 in development
 --------------
 
+* Support indentation of custom length (bug-fix)
+* Support tabs as indentation (new-feature)
+* Exclude params from ref (bug-fix)
 
 0.2.0 - August 14th, 2015 (alpha-1)
 -----------------------------------